Finish small refactor
