+++ CMakeLists.txt
@@ -18,7 +18,7 @@ set(LOADABLE_EXTENSION_NAME ${TARGET_NAME}_loadable_extension)
 project(${TARGET_NAME})
 include_directories(src/include)
 
-set(EXTENSION_SOURCES src/nanoarrow_extension.cpp src/read_arrow.cpp
+set(EXTENSION_SOURCES src/ipc/array_stream.cpp src/ipc/stream_factory.cpp src/ipc/stream_reader.cpp src/nanoarrow_extension.cpp src/read_arrow.cpp
                       src/write_arrow_stream.cpp)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
