+++ src/include/ipc/stream_factory.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB - NanoArrow
 //
-// ipc_stream_factory.hpp
+// ipc/ipc_stream_factory.hpp
 //
 //
 //===----------------------------------------------------------------------===//
@@ -9,6 +9,9 @@
 #pragma once
 
 #include "ipc/array_stream.hpp"
+
+#include "duckdb/common/arrow/arrow_wrapper.hpp"
+#include "duckdb/function/table/arrow.hpp"
 
 namespace duckdb {
 namespace ext_nanoarrow {
@@ -20,51 +23,17 @@ namespace ext_nanoarrow {
 class ArrowIpcArrowArrayStreamFactory {
 public:
   explicit ArrowIpcArrowArrayStreamFactory(ClientContext& context,
-                                           const std::string& src_string)
-      : fs(FileSystem::GetFileSystem(context)),
-        allocator(BufferAllocator::Get(context)),
-        src_string(src_string) {};
+                                           std::string  src_string);
 
-  // Called once when initializing Scan States
-  static unique_ptr<ArrowArrayStreamWrapper> Produce(uintptr_t factory_ptr,
-                                                     ArrowStreamParameters& parameters) {
-    auto factory = static_cast<ArrowIpcArrowArrayStreamFactory*>(
-        reinterpret_cast<void*>(factory_ptr));
+  //! Called once when initializing Scan States
+  static unique_ptr<ArrowArrayStreamWrapper> Produce(uintptr_t factory_ptr,  ArrowStreamParameters& parameters);
 
-    if (!factory->reader) {
-      throw InternalException("IpcStreamReader was not initialized or was already moved");
-    }
+  //! Get the schema of the arrow object
+  void GetFileSchema(ArrowSchemaWrapper& schema) const;
 
-    if (!parameters.projected_columns.columns.empty()) {
-      factory->reader->SetColumnProjection(parameters.projected_columns.columns);
-    }
-
-    auto out = make_uniq<ArrowArrayStreamWrapper>();
-    IpcArrayStream(std::move(factory->reader)).ToArrayStream(&out->arrow_array_stream);
-    return out;
-  }
-
-  // Get the schema of the arrow object
-  void GetFileSchema(ArrowSchemaWrapper& schema) {
-    if (!reader) {
-      throw InternalException("IpcStreamReader is no longer valid");
-    }
-
-    NANOARROW_THROW_NOT_OK(
-        ArrowSchemaDeepCopy(reader->GetFileSchema(), &schema.arrow_schema));
-  }
-
-  // Opens the file, wraps it in the ArrowIpcInputStream, and wraps it in
-  // the ArrowArrayStream reader.
-  void InitReader() {
-    if (reader) {
-      throw InternalException("ArrowArrayStream or IpcStreamReader already initialized");
-    }
-
-    unique_ptr<FileHandle> handle =
-        fs.OpenFile(src_string, FileOpenFlags::FILE_FLAGS_READ);
-    reader = make_uniq<IpcStreamReader>(fs, std::move(handle), allocator);
-  }
+  //! Opens the file, wraps it in the ArrowIpcInputStream, and wraps it in
+  //! the ArrowArrayStream reader.
+  void InitReader();
 
   FileSystem& fs;
   Allocator& allocator;
